BUGFIX: Make tests independent of DBAL
The tests should not rely on doctrine/dbal as
it is not a dependency of this package. The condition is
triggered for any exception that is not specifically
`\Neos\RedirectHandler\Exception` therefore we can just
use a generic `\Exception` in the test.

With dual compatibility for DBAL 2.x and 3.x in Neos 8.3
and up we cannot be sure the DBALException class exists in
an installation, therefore it now becomes an issue.